Cleanup the code
