+++ src/modules/espresso/Espresso.Shell/Models/EspressoSettingsModel.cs
@@ -1,4 +1,8 @@
-﻿using Newtonsoft.Json;
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Newtonsoft.Json;
 
 namespace Espresso.Shell.Models
 {
